% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toinfo.R
\name{csv2info}
\alias{csv2info}
\title{CSV file to Data Dictionary}
\usage{
csv2info(file, ...)
}
\arguments{
\item{file}{character string naming the file to read with \code{\link[utils]{read.csv}}}

\item{\dots}{Additional arguments, passed to \code{\link[utils]{read.csv}}.}
}
\value{
list
}
\description{
Convert CSV file to a data dictionary
}
\examples{
dd <- csv2info(system.file("examples", "ex_script.csv", package = "rcmoonpie"))
dd2df(dd)

}
