% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadictto.R
\name{dd2csv}
\alias{dd2csv}
\title{Data Dictionary to CSV file}
\usage{
dd2csv(l, file = "", ...)
}
\arguments{
\item{l}{data dictionary list}

\item{file}{character string naming the file to print to; defaults to standard output}

\item{\dots}{Additional arguments, passed to \code{\link[utils]{write.csv}}.}
}
\value{
creates CSV file
}
\description{
Convert data dictionary to a CSV file
}
\examples{
dd <- script2info(system.file("examples", "ex_script.R", package = "rcmoonpie"))
dd2csv(dd)

}
