% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadictto.R
\name{dd2df}
\alias{dd2df}
\title{Data Dictionary to data.frame}
\usage{
dd2df(l)
}
\arguments{
\item{l}{data dictionary list}
}
\value{
data.frame
}
\description{
Convert data dictionary to a data.frame
}
\examples{
dd <- script2info(system.file("examples", "ex_script.R", package = "rcmoonpie"))
dd2df(dd)

}
