% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toinfo.R
\name{df2info}
\alias{df2info}
\title{data.frame to Data Dictionary}
\usage{
df2info(df)
}
\arguments{
\item{df}{data.frame with variable meta-data}
}
\value{
list
}
\description{
Convert data.frame to data dictionary
}
\examples{
dd <- script2info(system.file("examples", "ex_script.R", package = "rcmoonpie"))
df <- dd2df(dd)
dd_alt <- df2info(df)
all.equal(dd, dd_alt)

}
