% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude.R
\name{excludeLevel}
\alias{excludeLevel}
\title{Exclude Factor Level}
\usage{
excludeLevel(l, dataset, variable, level, true = TRUE)
}
\arguments{
\item{l}{data dictionary list}

\item{dataset}{name of data set}

\item{variable}{name of factor variable}

\item{level}{name of factor level to remove}

\item{true}{The value to assign to \sQuote{exclude}, assumed to be TRUE.}
}
\value{
updated data dictionary list
}
\description{
Remove a level for a factor variable in the data set
}
\examples{
dd <- script2info(system.file("examples", "ex_script.R", package = "rcmoonpie"))
dd <- excludeLevel(dd, 'data', 'sex', 666)
dd$dataset[[1]]$variables[[5]]$factor[[3]]$exclude

}
