% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude.R
\name{excludeVar}
\alias{excludeVar}
\title{Exclude Variable}
\usage{
excludeVar(l, dataset, variable, true = TRUE)
}
\arguments{
\item{l}{data dictionary list}

\item{dataset}{name of data set}

\item{variable}{name of variable to remove}

\item{true}{The value to assign to \sQuote{exclude}, assumed to be TRUE.}
}
\value{
updated data dictionary list
}
\description{
Remove a variable from the data set.
}
\examples{
dd <- script2info(system.file("examples", "ex_script.R", package = "rcmoonpie"))
dd <- excludeVar(dd, 'data', 'redcap_event_name')
dd$dataset[[1]]$variables[[2]]$exclude

}
