% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vflist.R
\name{factorList}
\alias{factorList}
\title{Factor Variable Information}
\usage{
factorList(l)
}
\arguments{
\item{l}{data dictionary list}
}
\value{
data.frame
}
\description{
Data.frame with meta-data for factor variables
}
\examples{

}
