% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{findDataset}
\alias{findDataset}
\title{Find Dataset Within Data Dictionary List}
\usage{
findDataset(l, dataset)
}
\arguments{
\item{l}{data dictionary list}

\item{dataset}{name of data set}
}
\value{
integer
}
\description{
Find index location for a named data set within
the data dictionary.
}
\examples{
dd <- script2info(system.file("examples", "ex_script.R", package = "rcmoonpie"))
findDataset(dd, 'data')

}
