% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{findVariable}
\alias{findVariable}
\title{Find Variable Within Data Dictionary List}
\usage{
findVariable(l, dataset, variable)
}
\arguments{
\item{l}{data dictionary list}

\item{dataset}{name of data set}

\item{variable}{name of variable to remove}
}
\value{
list with data set location and variable location
}
\description{
Find index location for a named variable within
the data dictionary.
}
\examples{
dd <- script2info(system.file("examples", "ex_script.R", package = "rcmoonpie"))
findVariable(dd, 'data', 'redcap_event_name')

}
