% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{findVariableByPattern}
\alias{findVariableByPattern}
\title{Find Variable By Pattern}
\usage{
findVariableByPattern(l, pattern)
}
\arguments{
\item{l}{data dictionary list}

\item{pattern}{character value passed to \code{grep} function}
}
\value{
character vector of variable names
}
\description{
Find variable names that match a pattern.
}
\examples{
dd <- script2info(system.file("examples", "ex_script.R", package = "rcmoonpie"))
findVariableByPattern(dd, 'redcap')
# regular expression for strings that start with \dQuote{s}
findVariableByPattern(dd, '^s')

}
