% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcmoonpie-package.R
\docType{package}
\name{rcmoonpie-package}
\alias{rcmoonpie}
\alias{rcmoonpie-package}
\title{RedCap Moonpie}
\description{
add details
}
\details{
This package is experimental. Please submit bugs to
\url{https://github.com/couthcommander/rcmoonpie}.
}
\examples{
dd <- script2info(system.file("examples", "ex_script.R", package = "rcmoonpie"))
# exclude a variable
dd <- excludeVar(dd, 'data', 'redcap_event_name')
# convert to data.frame
dd2df(dd)
# reproduce R script
dd2script(dd, factorHandle = 'unchanged')
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/couthcommander/rcmoonpie}
}

}
\author{
Cole Beck \email{cole.beck@vumc.org}

Maintainer: Cole Beck \email{cole.beck@vumc.org}
}
