% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toinfo.R
\name{script2info}
\alias{script2info}
\title{R script to Data Dictionary}
\usage{
script2info(file, txt = NULL)
}
\arguments{
\item{file}{character string naming the file to read with \code{\link{scan}}}

\item{txt}{character string used as an alternative to reading}
}
\value{
list
}
\description{
Convert R script to a data dictionary
}
\examples{
dd <- script2info(system.file("examples", "ex_script.R", package = "rcmoonpie"))
dd2df(dd)

}
