% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude.R
\name{unexcludeLevel}
\alias{unexcludeLevel}
\title{Include Factor Level}
\usage{
unexcludeLevel(l, dataset, variable, level)
}
\arguments{
\item{l}{data dictionary list}

\item{dataset}{name of data set}

\item{variable}{name of factor variable}

\item{level}{name of factor level to include}
}
\value{
updated data dictionary list
}
\description{
Include a level for a factor variable in the data set, by allowing a previously removed level.
}
\examples{
dd <- script2info(system.file("examples", "ex_script.R", package = "rcmoonpie"))
dd <- excludeLevel(dd, 'data', 'sex', 666)
dd <- unexcludeLevel(dd, 'data', 'sex', 666)
dd$dataset[[1]]$variables[[5]]$factor[[3]]$exclude

}
