% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude.R
\name{unexcludeVar}
\alias{unexcludeVar}
\title{Include Variable}
\usage{
unexcludeVar(l, dataset, variable)
}
\arguments{
\item{l}{data dictionary list}

\item{dataset}{name of data set}

\item{variable}{name of variable to include}
}
\value{
updated data dictionary list
}
\description{
Include a variable in the data set, by allowing a previously removed variable.
}
\examples{
dd <- script2info(system.file("examples", "ex_script.R", package = "rcmoonpie"))
dd <- excludeVar(dd, 'data', 'redcap_event_name')
dd <- unexcludeVar(dd, 'data', 'redcap_event_name')
dd$dataset[[1]]$variables[[2]]$exclude

}
