% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vflist.R
\name{variableList}
\alias{variableList}
\title{Variable Information}
\usage{
variableList(l)
}
\arguments{
\item{l}{data dictionary list}
}
\value{
data.frame
}
\description{
Data.frame with variable meta-data
}
\examples{

}
