% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toinfo.R
\name{yaml2info}
\alias{yaml2info}
\title{YAML file to Data Dictionary}
\usage{
yaml2info(file)
}
\arguments{
\item{file}{character string naming the file to read with \code{\link[yaml]{yaml.load_file}}}
}
\value{
list
}
\description{
Convert YAML file to a data dictionary
}
\examples{
dd <- yaml2info(system.file("examples", "ex_script.yaml", package = "rcmoonpie"))
dd2df(dd)

}
